+++ lib/ig_markets/model/typecasters.rb
@@ -1,0 +1,78 @@
+module IGMarkets
+  # Implement typecaster methods for Model.
+  class Model
+    class << self
+      private
+
+      def typecaster_for(type)
+        if [Boolean, String, Fixnum, Float, Symbol, Date, Time].include? type
+          method "typecaster_#{type.to_s.gsub(/\AIGMarkets::/, '').downcase}"
+        elsif type.respond_to? :from
+          -> (value, _options) { type.from value }
+        end
+      end
+
+      def typecaster_boolean(value, _options)
+        return value if [nil, true, false].include? value
+
+        raise ArgumentError, "#{self}: invalid boolean value: #{value}"
+      end
+
+      def typecaster_string(value, options)
+        return nil if value.nil?
+
+        if options.key? :regex
+          raise ArgumentError, "#{self}: invalid string value: #{value}" unless options[:regex].match value.to_s
+        end
+
+        value.to_s
+      end
+
+      def typecaster_fixnum(value, _options)
+        return nil if value.nil?
+
+        value.to_s.to_i
+      end
+
+      def typecaster_float(value, _options)
+        return nil if value.nil?
+
+        Float(value)
+      end
+
+      def typecaster_symbol(value, _options)
+        return nil if value.nil?
+
+        value.to_s.downcase.to_sym
+      end
+
+      def typecaster_date(value, options)
+        raise ArgumentError, "#{self}: invalid or missing date format" unless options[:format].is_a? String
+
+        if value.is_a? String
+          begin
+            Date.strptime value, options[:format]
+          rescue ArgumentError
+            raise ArgumentError, "#{self}: failed parsing date '#{value}' with format '#{options[:format]}'"
+          end
+        else
+          value
+        end
+      end
+
+      def typecaster_time(value, options)
+        raise ArgumentError, "#{self}: invalid or missing time format" unless options[:format].is_a? String
+
+        if value.is_a?(String) || value.is_a?(Fixnum)
+          begin
+            Time.strptime "#{value}#{options[:time_zone]}", "#{options[:format]}#{'%z' if options[:time_zone]}"
+          rescue ArgumentError
+            raise ArgumentError, "#{self}: failed parsing time '#{value}' with format '#{options[:format]}'"
+          end
+        else
+          value
+        end
+      end
+    end
+  end
+end