+++ spec/factories/ig_markets/instrument_expiry_details.rb
@@ -1,6 +1,6 @@
 FactoryGirl.define do
   factory :instrument_expiry_details, class: IGMarkets::Instrument::ExpiryDetails do
-    last_dealing_date '2015/01/20 12:13:14'
+    last_dealing_date '2022-12-20T23:59'
     settlement_info 'info'
   end
 end