+++ spec/ig_markets/model_spec.rb
@@ -3,7 +3,8 @@ describe IGMarkets::Model do
     attribute :id
     attribute :bool, IGMarkets::Boolean
     attribute :string, String, regex: /\A[A-Z]{3}\Z/, nil_if: '-'
-    attribute :date, DateTime, format: '%Y-%m-%d'
+    attribute :date, Date, format: '%F'
+    attribute :time, Time, format: '%F %T', time_zone: '+0630'
     attribute :float, Float
     attribute :symbol, Symbol, allowed_values: [:a, :b]
   end
@@ -12,7 +13,7 @@ describe IGMarkets::Model do
 
   it 'initializes with specified attribute values' do
     expect(TestModel.new(id: 'test', bool: true).attributes).to eq(
-      id: 'test', bool: true, string: nil, date: nil, float: nil, symbol: nil)
+      id: 'test', bool: true, string: nil, date: nil, time: nil, float: nil, symbol: nil)
   end
 
   it 'fails when initialized with an unknown attribute' do
@@ -20,17 +21,19 @@ describe IGMarkets::Model do
   end
 
   it 'has the correct getter and setter methods' do
-    [:id, :id=, :bool, :bool=, :string, :string=, :date, :date=, :float, :float=, :symbol, :symbol=].each do |id|
+    [:id, :id=, :bool, :bool=, :string, :string=, :date, :date=, :time, :time=,
+     :float, :float=, :symbol, :symbol=].each do |id|
       expect(model.respond_to?(id)).to eq(true)
     end
   end
 
   it 'has the correct attributes hash' do
-    expect(model.attributes).to eq(id: nil, bool: nil, string: nil, date: nil, float: nil, symbol: nil)
+    expect(model.attributes).to eq(id: nil, bool: nil, string: nil, date: nil, time: nil, float: nil, symbol: nil)
   end
 
   it 'inspects attributes' do
-    expect(model.inspect).to eq('#<TestModel id: nil, bool: nil, string: nil, date: nil, float: nil, symbol: nil>')
+    expect(model.inspect).to eq(
+      '#<TestModel id: nil, bool: nil, string: nil, date: nil, time: nil, float: nil, symbol: nil>')
   end
 
   it 'inspects attributes in nested models' do
@@ -41,7 +44,7 @@ describe IGMarkets::Model do
     model = TestModel2.new test: TestModel.new
 
     expect(model.inspect).to eq(
-      '#<TestModel2 test: #<TestModel id: nil, bool: nil, string: nil, date: nil, float: nil, symbol: nil>>')
+      '#<TestModel2 test: #<TestModel id: nil, bool: nil, string: nil, date: nil, time: nil, float: nil, symbol: nil>>')
   end
 
   it '#from accepts nil' do
@@ -50,7 +53,7 @@ describe IGMarkets::Model do
 
   it '#from accepts an attributes hash' do
     expect(TestModel.from(id: 'test').attributes).to eq(
-      id: 'test', bool: nil, string: nil, date: nil, float: nil, symbol: nil)
+      id: 'test', bool: nil, string: nil, date: nil, time: nil, float: nil, symbol: nil)
   end
 
   it '#from accepts an instance and creates a copy' do
@@ -65,7 +68,7 @@ describe IGMarkets::Model do
   end
 
   it '#from raises on invalid inputs' do
-    ['', DateTime.new, IGMarkets::Model.new].each do |invalid_input|
+    ['', Time.new, IGMarkets::Model.new].each do |invalid_input|
       expect { TestModel.from(invalid_input) }.to raise_error(ArgumentError)
     end
   end
@@ -79,7 +82,15 @@ describe IGMarkets::Model do
   end
 
   it 'raises ArgumentError for an invalid date' do
-    expect { model.date = '2015-29-01' }.to raise_error(ArgumentError)
+    expect { model.date = '2015-29-01 A:B:C' }.to raise_error(ArgumentError)
+  end
+
+  it 'raises ArgumentError for an invalid time' do
+    expect { model.time = '2015-29-01 09:30:40' }.to raise_error(ArgumentError)
+  end
+
+  it 'raises ArgumentError for an invalid time' do
+    expect { model.time = '2015-29-01' }.to raise_error(ArgumentError)
   end
 
   it 'raises ArgumentError for an invalid float' do
@@ -101,7 +112,12 @@ describe IGMarkets::Model do
 
   it 'correctly parses a date in the expected format' do
     model.date = '2015-01-10'
-    expect(model.date).to eq(DateTime.new(2015, 1, 10))
+    expect(model.date).to eq(Date.new(2015, 1, 10))
+  end
+
+  it 'correctly parses a time in the expected format' do
+    model.time = '2015-01-10 09:30:25'
+    expect(model.time).to eq(Time.new(2015, 1, 10, 9, 30, 25, '+06:30'))
   end
 
   context 'with all attributes set' do
@@ -109,19 +125,22 @@ describe IGMarkets::Model do
       model.id = 'id'
       model.bool = true
       model.string = 'ABC'
-      model.date = '2015-01-10'
+      model.date = '2015-01-09'
+      model.time = '2015-01-10 00:00:00'
       model.float = '1.0'
       model.symbol = 'a'
     end
 
     it 'has the correct attributes hash' do
       expect(model.attributes).to eq(
-        id: 'id', bool: true, string: 'ABC', date: DateTime.new(2015, 1, 10), float: 1.0, symbol: :a)
+        id: 'id', bool: true, string: 'ABC', date: Date.new(2015, 1, 9), time: Time.new(2015, 1, 10, 0, 0, 0, '+06:30'),
+        float: 1.0, symbol: :a)
     end
 
     it 'inspects attributes' do
       expect(model.inspect).to eq(
-        '#<TestModel id: "id", bool: true, string: "ABC", date: 2015-01-10T00:00:00+00:00, float: 1.0, symbol: :a>')
+        '#<TestModel id: "id", bool: true, string: "ABC", date: 2015-01-09, time: 2015-01-10 00:00:00 +0630, ' \
+        'float: 1.0, symbol: :a>')
     end
   end
 end