+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 ### 0.2.pre - Unreleased
 
-- Added `IGMarkets::SprintMarketPosition#formatted_size`
+- `IGMarkets::Model` now has separate `Date` and `Time` attribute types, and a new `:time_zone` option is used for
+  `Time` attributes that have a known time zone. Previous uses of `DateTime` should be replaced with either `Date` or
+  `Time`.
+- Changed `size` attribute to always be of type `Float` on all models.
+- Changed `limit_distance` and `stop_distance` attributes to always be of type `Fixnum` on all models.
+- Added `IGMarkets::Format` module
+- Added `#expired?` and `#seconds_till_expiry` to `IGMarkets::SprintMarketPosition`
 - Fixed `IGMarkets::RequestFailedError#message`
 
 ### 0.1 - April 8, 2016